+++ romancal/outlier_detection/tests/test_outlier_detection.py
@@ -1,5 +1,6 @@
 import os
 
+import astropy.units as u
 import numpy as np
 import pytest
 from astropy.units import Quantity
@@ -217,6 +218,7 @@ def test_find_outliers(tmp_path, base_image):
         img.err[:] = err_value
         img.meta.filename = f"img{i}_suffix.asdf"
         img.meta.observation.exposure = i
+        img.meta.background.level = 0 * u.DN / u.s
         imgs.append(img)
 
     # add outliers
@@ -257,6 +259,7 @@ def test_identical_images(tmp_path, base_image, caplog):
     """
     img_1 = base_image()
     img_1.meta.filename = "img1_suffix.asdf"
+    img_1.meta.background.level = 0 * u.DN / u.s
     # add outliers
     img_1_input_coords = np.array(
         [(5, 45), (25, 25), (45, 85), (65, 65), (85, 5)], dtype=[("x", int), ("y", int)]
@@ -312,8 +315,10 @@ def test_outlier_detection_always_returns_modelcontainer_with_updated_datamodels
     os.chdir(tmp_path)
     img_1 = base_image()
     img_1.meta.filename = "img_1.asdf"
+    img_1.data *= img_1.meta.photometry.conversion_megajanskys / img_1.data.unit
     img_2 = base_image()
     img_2.meta.filename = "img_2.asdf"
+    img_2.data *= img_2.meta.photometry.conversion_megajanskys / img_2.data.unit
 
     library = ModelLibrary([img_1, img_2])
     library._save(tmp_path)