+++ CHANGES.rst
@@ -29,8 +29,11 @@ general
 
 - Replace ModelContainer with ModelLibrary [#1241]
 
+- Updated sky background usage in code and tests to use maker utilities. [#1351]
+
 - Refactor general step input handling to avoid early closing of
   input files to allow using more lazy loading [#1342]
+
 
 
 source_catalog